+++ src/lib.rs
@@ -46,6 +46,8 @@ extern crate doc_comment;
 doc_comment::doctest!("../README.md");
 
 pub use scip_sys as ffi;
+pub mod branchrule;
+pub mod pricer;
 pub mod constraint;
 pub mod model;
 pub mod retcode;