+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 ### Added
+- Solving stats methods for number of nodes, time, and number of lp iterations. 
+- Branching rule plugin. 
+- Variable pricer plugin. 
+- Set partitioning constraints. 
 ### Fixed
 ### Changed
 ### Removed